#322 refac(fe): workspace entities features 분리
