+++ src/frontend/apps/web/src/entities/index.ts
@@ -1,4 +1,8 @@
 export * from './alarm';
 export * from './auth';
-export * from './user';
+export * from './chat';
 export * from './file-upload';
+export * from './stock';
+export * from './user';
+export * from './video';
+export * from './workspace';
