+++ src/frontend/apps/web/src/entities/workspace/index.ts
@@ -1,0 +1,1 @@
+export * from './model';
