+++ src/frontend/apps/web/src/entities/workspace/model/WorkspaceForm.schema.ts
