+++ src/frontend/apps/web/src/entities/workspace/model/WorkspaceId.ts
