+++ src/frontend/apps/web/src/entities/workspace/model/index.ts
@@ -1,0 +1,7 @@
+export type {
+  WorkspaceSubscriptionResponse,
+  UnreadSubscriptionResponse,
+} from './workspace-type';
+export { WorkspaceFormSchema } from './WorkspaceForm.schema';
+export { WORKSPACE_ID } from './WorkspaceId';
+export { getWorkspaceId } from './getWorkspaceId';
