+++ src/frontend/apps/web/src/entities/workspace/model/workspace-type.d.ts
