--- src/frontend/apps/web/src/features/workspace/lib/index.ts
@@ -1,1 +1,0 @@
-export { getWorkspaceId } from './get-workspace-id';
