+++ src/frontend/apps/web/src/features/workspace/model/index.ts
@@ -1,11 +1,5 @@
-export { WORKSPACE_ID } from './workspace-id';
-export { WorkspaceFormSchema } from './workspace-form-schema';
-export type {
-  WorkspaceSubscriptionResponse,
-  UnreadSubscriptionResponse,
-} from './subscription.type';
-export { useWorkspaceMessages } from './use-workspace-messages';
-export { useWorkspaceSubscription } from './use-workspace-subscription';
-export { useWorkspaceChannels } from './use-workspace-channels';
-export { useUnreadMessages } from './use-unread-messages';
-export { useUnreadSubscription } from './use-unread-subscription';
+export { useWorkspaceMessages } from './useWorkspaceMessages';
+export { useWorkspaceSubscription } from './useWorkspaceSubscription';
+export { useWorkspaceChannels } from './useWorkspaceChannels';
+export { useUnreadMessages } from './useUnreadMessages';
+export { useUnreadSubscription } from './useUnreadSubscription';
